GTA-536

+++ GTA/Communication/SessionExpiredHandler.swift
@@ -15,7 +15,7 @@ class SessionExpiredHandler: ExpiredSessionDelegate {
         if let _ = data {
             metaData = try? DataParser.parse(data: data!)
         }
-        guard let meta = metaData, meta.meta?.userMessage == "Session expired", let _ = KeychainManager.getToken() else { return }
+        guard let meta = metaData, (meta.meta?.userMessage == "Session expired" || meta.meta?.userMessage == "Authorization error"), let _ = KeychainManager.getToken() else { return }
         print("Session expired")
         KeychainManager.deleteUsername()
         KeychainManager.deleteToken()
